+++ src/user-management/services/user.service.ts
@@ -1,4 +1,4 @@
-import { Injectable, BadRequestException } from '@nestjs/common';
+import { Injectable, BadRequestException, NotFoundException } from '@nestjs/common';
 import { InjectModel } from '@nestjs/mongoose';
 import { Model, Types } from 'mongoose';
 import { User, UserDocument } from '../schemas/user.schema';
@@ -7,6 +7,8 @@ import { CreateUserDto } from '../dto/create-user.dto';
 
 @Injectable()
 export class UserService {
+  
+  existsQuery: any = { deleted: false };
   constructor(@InjectModel(User.name) private userModel: Model<UserDocument>) {}
 
   async createUser(user: CreateUserDto): Promise<Types.ObjectId> {
@@ -24,7 +26,7 @@ export class UserService {
     role?: string,
     search?: string
   ): Promise<{ users: User[]; total: number }> {
-    const query = this.userModel.find();
+    const query = this.userModel.find(this.existsQuery);
 
     // Filtering by role
     if (role) {
@@ -55,18 +57,47 @@ export class UserService {
   }
 
   async findOne(id: string): Promise<User> {
+    // First check if the record exists and is not deleted
+    const existingRecord = await this.userModel.findOne({ _id: id, ...this.existsQuery }).exec();
+    if (!existingRecord) {
+      throw new NotFoundException('User not found or has been deleted.');
+    }
+
+    // Perform the update if the record is not marked as deleted
     return this.userModel.findById(id).exec();
   }
 
   async findOneByEmail(email: string): Promise<User | undefined> {
+    // First check if the record exists and is not deleted
+    const existingRecord = await this.userModel.findOne({ email, ...this.existsQuery }).exec();
+    if (!existingRecord) {
+      throw new NotFoundException('User not found or has been deleted.');
+    }
+
+    // Perform the update if the record is not marked as deleted
     return this.userModel.findOne({ email }).exec();
   }  
 
   async update(id: string, user: User): Promise<User> {
+    
+    // First check if the record exists and is not deleted
+    const existingRecord = await this.userModel.findOne({ _id: id, ...this.existsQuery }).exec();
+    if (!existingRecord) {
+      throw new NotFoundException('User not found or has been deleted.');
+    }
+
+    // Perform the update if the record is not marked as deleted
     return this.userModel.findByIdAndUpdate(id, user, { new: true }).exec();
   }
 
   async remove(id: string): Promise<User> {
+    // First check if the record exists and is not deleted
+    const existingRecord = await this.userModel.findOne({ _id: id, ...this.existsQuery }).exec();
+    if (!existingRecord) {
+      throw new NotFoundException('User not found or has been deleted.');
+    }
+
+    // Perform the update if the record is not marked as deleted
     return this.userModel.findByIdAndRemove(id).exec();
   }
 
@@ -74,7 +105,7 @@ export class UserService {
     verificationKey: string,
     newPassword: string,
   ): Promise<Types.ObjectId> {
-    const user = await this.userModel.findOne({ verificationKey });
+    const user = await this.userModel.findOne({ verificationKey, ...this.existsQuery });
 
     if (!user) {
       throw new Error('User not found');
